requirements.txt
pyperclip
cryptography
mysql-connector-python
bcrypt
flask
